module4-task1 исправляет роуты для offer-page, переносит getRatingWidth в отдельный файл, исправляет id для моков
