+++ project/src/components/comment-form/comment-form.tsx
@@ -2,7 +2,7 @@ import React, { useState, ChangeEvent, FormEvent } from 'react';
 
 function CommentForm(): JSX.Element {
   const [formReviews, setFormReviews] = useState({
-    comment: '',
+    review: '',
     rating: 0,
   });
 
@@ -55,7 +55,14 @@ function CommentForm(): JSX.Element {
           </svg>
         </label>
       </div>
-      <textarea className="reviews__textarea form__textarea" id="review" name="review" placeholder="Tell how was your stay, what you like and what can be improved" />
+      <textarea
+        className="reviews__textarea form__textarea"
+        id="review"
+        name="review"
+        placeholder="Tell how was your stay, what you like and what can be improved"
+        onChange={ handleInputChange }
+        value={ formReviews.review }
+      />
       <div className="reviews__button-wrapper">
         <p className="reviews__help">
           To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with at least <b className="reviews__text-amount">50 characters</b>.
