+++ project/src/router/router.ts
@@ -1,5 +1,5 @@
 export enum AppRoute {
   Main = '/',
-  Login = './login',
-  Offer = './offer',
+  Login = '/login',
+  Offer = '/offer',
 }
